finish 8/4/2017

+++ ECClaimSystem/Controllers/ECManagerController.cs
@@ -60,7 +60,7 @@ namespace ECClaimSystem.Controllers
             return View(lsECClaim);
         }
 
-        public ActionResult Details(int id)
+        public ActionResult Details(long id)
         {
             ECClaim ec = _ecClaimRepository.GetECClaimById(id);
             Models.ECClaim claim = new Models.ECClaim();
